Implement features required for winning tokens
Lots of changes. Too many to list here.

Basically I had to redo the grid positioning logic (again), and add
lots of stuff to make it so that when the tile sides got full, the
tile could then discard them and give its tokens to the winning player.

This all required a few structural changes, so the code is a mess right
now.

The plan is to do a massive refactor once the remaining game logic has
been ported over and runs correctly.