Initial multiplayer rewrite commit
Adds NetworkManager and Network components to the scene.

Removes all scripts because they were causing sync issues between server
and client. This means that the code will need to be mostly re-written.
With some code copied over, such as the positioning code.

I've tested the prefabs that are in the scene, and they do sync
correctly with the client matching the editor when manually moving them
around.